move share_xaxes and share_yaxes to plotutil
